--- app/src/02_modules/debug/headless/index.ts
@@ -1,2 +1,0 @@
-export { useSecretCode } from './useSecretCode';
-export { debugLog, getLogs, clearLogs, subscribeLogs, setVerboseLogging, type LogEntry } from './debugLog';
