+++ app/src/02_modules/debug/services/index.ts
@@ -1,0 +1,19 @@
+/**
+ * Debug Module Services
+ */
+
+// Debug log collection service
+export {
+  debugLog,
+  getLogs,
+  clearLogs,
+  subscribeLogs,
+  setVerboseLogging,
+  type LogEntry,
+} from './debugLogService';
+
+// Debug settings state service
+export { debugSettingsStateService } from './debugSettingsStateService';
+
+// Diagnostic service
+export { DiagnosticService } from './DiagnosticService';
